+++ Sources/Frontend/Commands/ScanBehavior.swift
@@ -63,7 +63,7 @@ final class ScanBehavior {
             results = try block(project)
             let interval = logger.beginInterval("result:output")
             let filteredResults = OutputDeclarationFilter().filter(results)
-            let output = try configuration.outputFormat.formatter.init().format(filteredResults)
+            let output = try configuration.outputFormat.formatter.init(configuration: configuration).format(filteredResults)
             logger.info("", canQuiet: true)
             logger.info(output, canQuiet: false)
             logger.endInterval(interval)