+++ Sources/PeripheryKit/Formatters/CheckstyleFormatter.swift
@@ -1,13 +1,21 @@
 import Foundation
 import Shared
+import SystemPackage
 
 final class CheckstyleFormatter: OutputFormatter {
+    let configuration: Configuration
+    lazy var currentFilePath: FilePath = { .current }()
+
+    init(configuration: Configuration) {
+        self.configuration = configuration
+    }
+
     func format(_ results: [ScanResult]) -> String {
         let parts = results.flatMap { describe($0, colored: false) }
         let xml = [
             "<?xml version=\"1.0\" encoding=\"utf-8\"?>\n<checkstyle version=\"4.3\">",
             parts
-                .group(by: { ($0.0.file.path.string).escapedForXML() })
+                .group(by: { outputPath($0.0).string.escapedForXML() })
                 .sorted(by: { $0.key < $1.key })
                 .map(generateForFile).joined(),
             "\n</checkstyle>"