+++ Sources/PeripheryKit/Formatters/XcodeFormatter.swift
@@ -1,7 +1,15 @@
 import Foundation
 import Shared
+import SystemPackage
 
 final class XcodeFormatter: OutputFormatter {
+    let configuration: Configuration
+    lazy var currentFilePath: FilePath = { .current }()
+
+    init(configuration: Configuration) {
+        self.configuration = configuration
+    }
+
     func format(_ results: [ScanResult]) throws -> String {
         guard results.count > 0 else {
             return colorize("* ", .boldGreen) + colorize("No unused code detected.", .bold)
@@ -18,14 +26,14 @@ final class XcodeFormatter: OutputFormatter {
     // MARK: - Private
 
     private func prefix(for location: SourceLocation) -> String {
-        let absPath = location.file.path.lexicallyNormalized()
-        let path = absPath.removingLastComponent().string
-        let file = colorize(absPath.lastComponent?.stem ?? "", .bold)
-        let ext = absPath.extension ?? "swift"
+        let path = outputPath(location)
+        let dir = path.removingLastComponent()
+        let file = colorize(path.lastComponent?.stem ?? "", .bold)
+        let ext = path.extension ?? "swift"
         let lineNum = colorize(String(location.line), .bold)
         let column = location.column
         let warning = colorize("warning:", .boldYellow)
 
-        return "\(path)/\(file).\(ext):\(lineNum):\(column): \(warning) "
+        return "\(dir)/\(file).\(ext):\(lineNum):\(column): \(warning) "
     }
 }